move note toolbar to a generic entity toolbar
